Published or Booked publication

+++ src/Publications/publications.service.ts
@@ -2,6 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { PublicationsRepository } from './publications.repository';
 import { CreatePublicationsDTO } from 'src/Dtos/publications.dtos';
 import { NotFoundException } from '@nestjs/common';
+import dayjs from 'dayjs';
+import { ForbiddenException } from '@nestjs/common';
 
 @Injectable()
 export class PublicationsService {
@@ -33,25 +35,31 @@ export class PublicationsService {
     }
   }
 
-  updatePublication(publication: CreatePublicationsDTO, id: number) {
+  async updatePublication(publication: CreatePublicationsDTO, id: number) {
     const media = this.publiRepository.getMediaById(publication.mediaId);
     const post = this.publiRepository.getPostById(publication.postId);
     if (!media || !post) {
       throw new NotFoundException('NOT FOUND');
     } else {
-      const update = this.publiRepository.updatePublication(id, publication);
-      if (update) {
-        return update;
+      const publication = await this.publiRepository.getPublicationById(id);
+      if (publication) {
+        const now = dayjs().toDate().getTime();
+        const timestamp = dayjs(publication.date).toDate().getTime();
+        if (now < timestamp) {
+          return this.publiRepository.updatePublication(id, publication);
+        } else {
+          throw new ForbiddenException('FORBIDDEN');
+        }
       } else {
         throw new NotFoundException('NOT FOUND');
       }
     }
   }
 
-  deletePublication(id: number) {
-    const del = this.publiRepository.deletePublication(id);
-    if (del) {
-      return del;
+  async deletePublication(id: number) {
+    const publication = this.publiRepository.getPublicationById(id);
+    if (publication) {
+      return this.publiRepository.deletePublication(id);
     } else {
       throw new NotFoundException('NOT FOUND');
     }
